+++ test/utils/filter/lzw_decode_filter_test.dart
@@ -1,4 +1,5 @@
 import 'dart:convert';
+import 'dart:typed_data';
 
 import 'package:dart_pdf_reader/src/model/pdf_types.dart';
 import 'package:dart_pdf_reader/src/utils/filter/stream_filter.dart';
@@ -9,7 +10,8 @@ void main() {
     group('LZWDecodeFilter tests', () {
       test('Decoding test', () {
         final decoded = StreamFilter(const PDFName('LZWDecode')).decode(
-          [0x80, 0x0B, 0x60, 0x50, 0x22, 0x0C, 0x0C, 0x85, 0x01],
+          Uint8List.fromList(
+              [0x80, 0x0B, 0x60, 0x50, 0x22, 0x0C, 0x0C, 0x85, 0x01]),
           null,
           const PDFDictionary({}),
         );